Refactor into a cloud-based project
