+++ Cargo.toml
@@ -12,7 +12,6 @@ serde = { version = "1.0.140", features = ["derive"] }
 ansi-to-html = { path = "ansi-to-html" }
 color-eyre = "0.6.2"
 rayon = "1.5.3"
-ring = "0.16"
 semver = "1.0.12"
 tar = "0.4.38"
 flate2 = "1.0.24"
@@ -20,7 +19,6 @@ ureq = "2.5.0"
 once_cell = "1.13.0"
 csv = "1.1.6"
 fxhash = "0.2.1"
-indicatif = "0.17.0"
 regex = "1.6.0"
 uuid = { version = "1.1.2", features = ["v4"] }
 num_cpus = "1.13.1"
