+++ docker/run-asan.sh
@@ -12,7 +12,6 @@ do
         cargo +nightly update --color=always
         cargo +nightly careful test --no-run --color=always $ARGS
         unbuffer -p /usr/bin/time -v cargo +nightly careful test --no-fail-fast -- --test-threads=2 $ARGS
-        cat Cargo.lock
     fi
     echo "-${TEST_END_DELIMITER}-"
 done < /dev/stdin
